Re-add fixed delay between tests and update idempotency key between retries
